Laba5.1

+++ Secure/Secure/FormCreateOrder.cs
@@ -79,7 +79,7 @@ namespace Secure
                 {
                     KomlectId = Convert.ToInt32(comboBoxKomlect.SelectedValue),
                     Count = Convert.ToInt32(textBoxCount.Text),
-                    Sum = Convert.ToDecimal(textBoxSum.Text)
+                    Sum = Convert.ToInt32(textBoxSum.Text)
                 });
                 MessageBox.Show("Сохранение прошло успешно", "Сообщение",
                MessageBoxButtons.OK, MessageBoxIcon.Information);
