+++ Secure/SecureShopFileImplement/Models/Order.cs
@@ -13,7 +13,7 @@ namespace SecureShopFileImplement.Models
         public int Id { get; set; }
         public int KomlectId { get; set; }
         public int Count { get; set; }
-        public decimal Sum { get; set; }
+        public int Sum { get; set; }
         public OrderStatus Status { get; set; }
         public DateTime DateCreate { get; set; }
         public DateTime? DateImplement { get; set; }
